fix: post-refactor fixes — polling error retention, nullable Uri, sessions type mismatch
- ProxyManager.polling: transition to ERROR not STOPPED on crash (was losing error)
- MainActivity: null-safe refreshFileInfo call (fileUri is nullable)
- ProxyState.sessions: Int → Long (gomobile int maps to Java long)

+++ android/app/src/main/java/com/flowdav/app/ProxyState.kt
@@ -2,7 +2,7 @@ package com.flowdav.app
 
 data class ProxyState(
     val running: Boolean = false,
-    val sessions: Int = 0,
+    val sessions: Long = 0L,
     val listenAddr: String = "",
     val logs: String = "",
     val error: String? = null,
